refactor: rename and enhance donation handling methods
Renamed `handleDonations` to `handleDonationsButton` for clarity and updated its usage across multiple components. Introduced methods for handling rich text items more flexibly while ensuring type safety and consistency. Added detailed margin and styling configurations, enhancing layout handling.

+++ lib/MBMigration/Builder/Layout/Common/Concern/SectionStylesAble.php
@@ -184,7 +184,7 @@ trait SectionStylesAble
                 $brizySection
                     ->getParent()
                     ->getValue()
-                    ->set_sectionHeight(str_replace('px', '', $sectionStyles['height']) ?? 500)
+                    ->set_sectionHeight((int) str_replace('px', '', $sectionStyles['height']) ?? 500)
                     ->set_fullHeight('custom');
             }
         } else if ($this->hasVideoBackground($mbSectionItem)) {
